Cleanup CMake boilerplate in top level makefile
This is not entirely satisfying because the IDF and the normal build system are
not, per-say, merged together. It's still a good start and remove much of the
boring stuff out of the top-level CMakeLists without breaking compatibility with
the IDF build system.

Unfortunately, it seems we can't do much override in the IDF build system
without changing the way the project is built. Trying another approach soon.