Arduino-ESP32 port
This adds an additional target, arduino_esp32, which builds the stack with the
Arduino library. On ESP32, this library is layed on top of the already existing
FreeRTOS code, which is why we can still use the NetworkInterface code nearly
as-is.

This patch obviously include what it takes for the library to be usable from a
C++ program but the amount of modifications needed for this to work has been
fairly limited in their scope.

Documentation about this target is to be added in the next commit.