Generate random ethernet address for remote side of Linux TAP
There was a bit of confusion there on what to do. When creating a tap interface,
Linux attributes it a random ethernet address. This address is actually NOT to
be used to send packets as it confuses Linux: packets coming out of an interface
should not have the same source address as the destination.

While this behaviour did not cause any problems in our tests, it was not
correct. I fixed this by generating an random ethernet address to act as the
remote side of the TAP.