Incoming enrollment is working between IRATI and RS
RS can now accept incoming enrollment from IRATI. This involved lots and lots of
changes to shoare up the fundation of the code for the future. Extensive work
was done in the neighbor SerDes.

At this point, the Enrollment component needs more work to deal with issues that
could happen during enrollment but the normal code path should work.