Fix fundamental memory leak in SerDes objects
Allocating a serObjectValue_t dynamically is problematic when we
want to use the serialized content of the CDAP message in a
netbuf. Freeing the netbuf means freeing the buffer that the
serObjectValue_t points too, which is not going to be the start
of the allocated buffer in our case. This fixes the problem by
making the allocation of the serObjectValue_t struct the problem
of the caller. The content and size of the serialized objects are
wrapped by the netbuf code.