Per-thread errors aren't platform specific
They are compiler specific! "__thread" is supported by
GCC and thus we do not need to do anything special for
the platforms we have to support since the only compiler
we use right now is GCC.

Moving 'rserr.c' to a directory called GCC makes this 
more obvious, but, for now, no real work is done in the 
CMakeLists.txt to enforce that.